requirements.txt
duckdb
pandas
pyyaml
minio
requests
flask
python-dotenv
firebase-admin